implemented Parent page

+++ src/app/pages/users/list.component.ts
@@ -1,25 +1,29 @@
-import { Component, OnInit } from '@angular/core';
-import { first } from 'rxjs/operators';
+import {Component, OnInit} from '@angular/core';
+import {first} from 'rxjs/operators';
 
-import { AccountService } from '../../services';
+import {AccountService} from '../../services';
+import {User} from "../../entities/user";
 
-@Component({ templateUrl: 'list.component.html' })
+@Component({templateUrl: 'list.component.html'})
 export class ListComponent implements OnInit {
-    users?: any[];
+  users?: User[];
 
-    constructor(private accountService: AccountService) {}
+  constructor(private accountService: AccountService) {
+  }
 
-    ngOnInit() {
-        this.accountService.getAll()
-            .pipe(first())
-            .subscribe(users => this.users = users);
+  ngOnInit() {
+    this.accountService.getAll()
+      .pipe(first())
+      .subscribe(users => this.users = users);
+  }
+
+  deleteUser(username: string) {
+    const user = this.users!.find(x => x.username === username);
+    if(user) {
+      user.isDeleting = true;
+      this.accountService.delete(username)
+        .pipe(first())
+        .subscribe(() => this.users = this.users!.filter(x => x.username !== username));
     }
-
-    deleteUser(username: string) {
-        const user = this.users!.find(x => x.username === username);
-        user.isDeleting = true;
-        this.accountService.delete(username)
-            .pipe(first())
-            .subscribe(() => this.users = this.users!.filter(x => x.username !== username));
-    }
+  }
 }
